# ai-services/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for managing AI Services catalog, applications, and authentication",
        "title": "AI Services Catalog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/project-ai-services/ai-services",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/architectures": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all available architecture templates",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "List available architectures",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.Architecture"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/architectures/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves detailed information about a specific architecture template",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Get architecture details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Architecture template ID (e.g., 'rag')",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.Architecture"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Architecture not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all deployable service templates. Dependency-only services are excluded from this list.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "List available services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ServiceSummary"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves detailed information about a specific service template",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Get service details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service template ID (e.g., 'summarize')",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ServiceSummary"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/{id}/params": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves custom parameters schema for a specific service template. Returns JSON Schema format that UI can use to generate dynamic forms with validation.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Get service custom parameters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service template ID (e.g., 'rag', 'summarize')",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON Schema for service parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.Architecture": {
            "type": "object",
            "properties": {
                "certified_by": {
                    "type": "string"
                },
                "code_link": {
                    "type": "string"
                },
                "demo_link": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "documentation_link": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ServiceReference"
                    }
                },
                "supported_runtimes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "\"architecture\"",
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DependencyReference": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ResourceRequirements": {
            "type": "object",
            "properties": {
                "min_cpu": {
                    "type": "string"
                },
                "min_disk": {
                    "type": "string"
                },
                "min_memory": {
                    "type": "string"
                },
                "min_spyre": {
                    "description": "Number of Spyre cards required",
                    "type": "integer"
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ServiceReference": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "optional": {
                    "type": "boolean"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ServiceSummary": {
            "type": "object",
            "properties": {
                "architectures": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "certified_by": {
                    "type": "string"
                },
                "dependencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DependencyReference"
                    }
                },
                "dependency_only": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "requirements": {
                    "description": "Runtime-specific resource requirements",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ResourceRequirements"
                        }
                    ]
                },
                "supported_runtimes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "internal_pkg_catalog_apiserver_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}